apps/backends/auth-service/requirements.txt
Django==4.2.*
psycopg2-binary>=2.9.10
redis==5.0.1
# django-redis==5.3.0
celery[redis]>=5.3.6
flower>=2.0.0
djangorestframework>=3.14.0
django-jazzmin==3.0.1
gunicorn>=21.2.0
uvicorn[standard]>=0.29.0
drf-spectacular[sidecar]>=0.27.2
whitenoise==6.8.2
django-cors-headers>=4.0.0
requests>=2.31.0
cryptography>=41.0.0
bcrypt>=4.0.0
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
django-import-export
django-timezone-field>=7.1
Pillow>=10.0.0
pika>=1.3.0